server: fix race condition for heartbeat
The heartbeater works on a timer, and is run in a goroutine to keep
track of the state of a connection.

The problem is though that the timer is not closed when the connection
is, and therefore, it may try to write into the connection outgoing
channel, making the whole application panic.

This commit fixes such a behaviour by ensuring the connection is not
closed when checking for the heartbeat.