server: mutex access to channels in getChannel
The channels of a connection is a shared map which is non-thread safe.

Because of this, accessing the map in Connection.getChannel is
inherently dangerous, and may fail with a concurrent access panic.

This commit adds a mutex lock before the access to the map so the crash
does not occur.