server: Signal readiness to systemd on Linux
This allows using systemd service files with Type=notify which means the
garagemq service will not be considered as started until it has signaled
that it's ready (after setting up the listening socket).
This causes any other service that has a dependency on garagemq to be
properly delayed until garagemq is up and running.

The daemon.go and daemon_linux.go are split up to avoid pulling
in the go-systemd dependency on non-linux.

Fixes #29